+++ Source/Nova/Audio/Midi.h
@@ -19,7 +19,8 @@ public:
     {
         EventType   eventType_ = {};
         int         noteNumber_ = 0;
-        float       time_ = 0.0f;
+        double       time_ = 0.0f;
+        //float       time_ = 0.0f;
         bool        judged_ = false; // 判定済みフラグを追加
     };
 
@@ -32,34 +33,38 @@ public:
 
     //  ノートを取得する関数
     std::vector<MidiNote> GetNotes() const { return notes_; }
-    const Midi::MidiNote* FindClosestNote(const float& inputTime);
-    Midi::MidiNote*       FindClosestNoteInLoop(const float& inputTime);
+    const Midi::MidiNote* FindClosestNote(const double& inputTime);
+    Midi::MidiNote*       FindClosestNoteInLoop(const double& inputTime);
     const Midi::MidiNote* GetNextNote(const float& currentTime);
 
     void	Initialize();								    //	初期化処理
-    void	Update(const float& elapsedTime);				//	更新処理
+    void	Update(const double& elapsedTime);				//	更新処理
+    //void	Update(const float& elapsedTime);				//	更新処理
     void	Finalize();									    //	終了処理
     void	DrawDebug();								    //	デバッグ描画
 
-    void UpdateCurrentTimer(const float& elapsedTime);      //  タイマー更新処理
+    void UpdateCurrentTimer(const double& elapsedTime);      //  タイマー更新処理
+    //void UpdateCurrentTimer(const float& elapsedTime);      //  タイマー更新処理
 
-    bool IsNoteOnAtTime(const float& time, const float& threshold = 0.17f);
-    bool IsInputNoteOn(const float& inputTime);             //  入力時間がノートオンかどうか
+    bool IsNoteOnAtTime(const double& time, const float& threshold = 0.17f);
+    bool IsInputNoteOn(const double& inputTime);             //  入力時間がノートオンかどうか
     bool IsCurrentTimeNoteOn();                             //  現在の時間がノートオンかどうか
 
     void    BuildNoteOnList();                              //  ノートオンリストを事前に構築
     void    SortNoteOnList();                              //  ノートオンリストをソート
-    float	GetNearMidiTime(const float& inputTime);        //	入力されたタイミングから近いノートを判定	
+    float	GetNearMidiTime(const double& inputTime);        //	入力されたタイミングから近いノートを判定	
 
     smf::MidiFile&  GetMidiFile()                   { return midiFile_; }	                //	midiファイル取得
     const float     GetMidiFileDurationSeconds()    { return midiFileDurationSeconds_; }    //  midiファイル全体の長さ取得
-    float           GetCurrentTimer() { return currentTimer_; }
+    double           GetCurrentTimer() { return currentTimer_; }
+    //float           GetCurrentTimer() { return currentTimer_; }
 
     void ResetJudgedNotes();    //  ノートの判定済みフラグをリセット
 
 private:
     float                   midiFileDurationSeconds_ = 0.0f;    //  midiファイル全体の長さ(時間[s])
-    float                   currentTimer_ = 0.0f;               //  現在の時間
+    double                   currentTimer_ = 0.0f;               //  現在の時間
+    //float                   currentTimer_ = 0.0f;               //  現在の時間
     smf::MidiFile	        midiFile_ = {};                     //  midiファイル
     std::vector<MidiNote>   notes_;                             //  midiデータ内のノート
